requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.16
factory-boy==2.12.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
python-dotenv==0.13.0
pytest==5.4.2
mysqlclient==1.4.6
